chore: renamed file to `opportunity-utils.js`
